restructured osb-broker UTs
