+++ broker/applications/osb-broker/test/acceptance/service-broker-api-2.0.instances.director.spec.js
@@ -2,7 +2,7 @@
 
 const _ = require('lodash');
 const Promise = require('bluebird');
-const app = require('../support/apps').internal;
+const app = require('../../../../test/test_broker/support/apps').internal;
 const config = require('@sf/app-config');
 const { catalog } = require('@sf/models');
 const ScheduleManager = require('@sf/jobs');
