+++ broker/applications/osb-broker/test/acceptance/service-broker-api.catalog.spec.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const _ = require('lodash');
-const app = require('../support/apps').internal;
+const app = require('../../../../test/test_broker/support/apps').internal;
 const { catalog } = require('@sf/models');
 const config = require('@sf/app-config');
 const { CONST } = require('@sf/common-utils');
