+++ broker/applications/osb-broker/test/acceptance/service-broker-api.instances.docker.spec.js
@@ -2,7 +2,7 @@
 
 const _ = require('lodash');
 const parseUrl = require('url').parse;
-const app = require('../support/apps').internal;
+const app = require('../../../../test/test_broker/support/apps').internal;
 const docker = require('@sf/docker');
 const config = require('@sf/app-config');
 const {
