+++ broker/applications/osb-broker/test/broker.middleware.spec.js
@@ -2,7 +2,7 @@
 
 const Promise = require('bluebird');
 const proxyquire = require('proxyquire');
-const middleware = proxyquire('../../applications/osb-broker/src/api-controllers/middleware', {
+const middleware = proxyquire('../src/api-controllers/middleware', {
   'basic-auth': function (req) {
     return req.auth;
   }
@@ -16,8 +16,8 @@ const {
 } = require('@sf/common-utils');
 const config = require('@sf/app-config');
 const { catalog } = require('@sf/models');
-const ServiceFabrikApiController = require('../../applications/extensions/src/api-controllers/ServiceFabrikApiController');
-const QuotaClient = require('../../applications/osb-broker/src/api-controllers/middleware/QuotaClient');
+const ServiceFabrikApiController = require('../../extensions/src/api-controllers/ServiceFabrikApiController');
+const QuotaClient = require('../src/api-controllers/middleware/QuotaClient');
 const PROMISE_WAIT_SIMULATED_DELAY = 30;
 
 class Response {
@@ -44,22 +44,22 @@ describe('#timeout', function () {
   before(function () {
     getInfoStub = sinon.stub(ServiceFabrikApiController.prototype, 'getInfo');
     config.http_timeout = 10;
-    delete require.cache[require.resolve('./support/apps')];
-    delete require.cache[require.resolve('../../applications/extensions/src/api-controllers/routes')];
-    delete require.cache[require.resolve('../../applications/extensions/src/api-controllers/routes/api')];
-    delete require.cache[require.resolve('../../applications/extensions/src/api-controllers/routes/api/v1')];
-    delete require.cache[require.resolve('../../applications/extensions/src/api-controllers')];
-    app = require('./support/apps').external;
+    delete require.cache[require.resolve('../../../test/test_broker/support/apps')];
+    delete require.cache[require.resolve('../../extensions/src/api-controllers/routes')];
+    delete require.cache[require.resolve('../../extensions/src/api-controllers/routes/api')];
+    delete require.cache[require.resolve('../../extensions/src/api-controllers/routes/api/v1')];
+    delete require.cache[require.resolve('../../extensions/src/api-controllers')];
+    app = require('../../../test/test_broker/support/apps').external;
   });
   after(function () {
     config.http_timeout = original_http_timeout;
     getInfoStub.restore();
-    delete require.cache[require.resolve('./support/apps')];
-    delete require.cache[require.resolve('../../applications/extensions/src/api-controllers/routes')];
-    delete require.cache[require.resolve('../../applications/extensions/src/api-controllers/routes/api')];
-    delete require.cache[require.resolve('../../applications/extensions/src/api-controllers/routes/api/v1')];
-    delete require.cache[require.resolve('../../applications/extensions/src/api-controllers')];
-    app = require('./support/apps').external;
+    delete require.cache[require.resolve('../../../test/test_broker/support/apps')];
+    delete require.cache[require.resolve('../../extensions/src/api-controllers/routes')];
+    delete require.cache[require.resolve('../../extensions/src/api-controllers/routes/api')];
+    delete require.cache[require.resolve('../../extensions/src/api-controllers/routes/api/v1')];
+    delete require.cache[require.resolve('../../extensions/src/api-controllers')];
+    app = require('../../../test/test_broker/support/apps').external;
   });
   it('should return 503 after timeout occurs', function () {
     return chai.request(app)
