+++ broker/test/init
@@ -2,12 +2,14 @@
 
 test_broker(){
     rm -f ./logs/test.log
+    rm -Rf coverage/
     case $1 in 
-        "test")             _mocha test/test_broker test/test_broker/acceptance --opts test/test_broker/mocha_broker.opts;;
-        "test-ci")          babel-node ./node_modules/.bin/isparta cover --report cobertura _mocha  -- test/test_broker/acceptance test/test_broker --opts test/test_broker/mocha_broker.opts;;
-        "test-coverall")    babel-node ./node_modules/.bin/isparta cover --report lcovonly _mocha -- test/test_broker/acceptance/ test/test_broker/ --opts test/test_broker/mocha_broker.opts;;
-        "test-cov")         babel-node ./node_modules/.bin/isparta cover _mocha -- test/test_broker test/test_broker/acceptance/ --opts test/test_broker/mocha_broker.opts;;
+        "test")             _mocha "test/test_broker/**/*.js" "applications/**/test/*.js" "applications/**/test/**/*.js" --opts test/test_broker/mocha_broker.opts;;
+        "test-ci")          nyc --reporter=cobertura --reporter=text-summary mocha "test/test_broker/**/*.js" "applications/**/test/*.js" "applications/**/test/**/*.js" --opts test/test_broker/mocha_broker.opts;;
+        "test-coverall")    nyc --reporter=lcov --reporter=text-summary mocha "test/test_broker/**/*.js" "applications/**/test/*.js" "applications/**/test/**/*.js" --opts test/test_broker/mocha_broker.opts && nyc report --reporter=text-lcov | coveralls;;
+        "test-cov")         nyc --reporter=lcov --reporter=text-summary mocha "test/test_broker/**/*.js" "applications/**/test/*.js" "applications/**/test/**/*.js" --opts test/test_broker/mocha_broker.opts;;
     esac
+    rm -Rf .nyc_output/
 }
 
 test_deployment_hooks(){
